+++ benchmarks/bench_ggsw_encrypt.cpp
@@ -4,7 +4,7 @@
 
 extern "C" {
 #include "bivariate_polynomial.h"
-#include "ggsw.h"
+#include "ggsw_arithmetic.h"
 #include "ggsw_ciphertext.h"
 #include "ggsw_params.h"
 #include "glwe_transform_key.h"
