+++ benchmarks/bench_glwegad_half_prod.cpp
@@ -4,9 +4,9 @@
 
 extern "C" {
 #include "bivariate_polynomial.h"
-#include "ggsw.h"
+#include "ggsw_arithmetic.h"
 #include "ggsw_params.h"
-#include "glwe.h"
+#include "glwe_arithmetic.h"
 #include "glwe_transform_key.h"
 #include "glwegadget.h"
 #include "glwegadget_ciphertext.h"
