+++ include/core/ggsw/ggsw_arithmetic.h
