+++ src/core/glwe/CMakeLists.txt
@@ -4,7 +4,7 @@
 
 # Source files for GLWE
 set(GLWE_SRCS
-    glwe.c
+    glwe_arithmetic.c
     glwe_params.c
     glwe_ciphertext.c
     glwe_key.c
