+++ tests/unit/core/ggsw/test_ggsw_encrypt.c
@@ -4,8 +4,8 @@
 #include <math.h>
 
 #include "bivariate_polynomial.h"
-#include "core/ggsw/ggsw.h"
-#include "core/glwe/glwe.h"
+#include "core/ggsw/ggsw_arithmetic.h"
+#include "core/glwe/glwe_arithmetic.h"
 #include "ggsw_ciphertext.h"
 #include "glwe_ciphertext.h"
 #include "glwe_key.h"
