+++ tests/unit/core/ggsw/test_glwegadget_encrypt.c
@@ -3,9 +3,9 @@
 #include <float.h>
 
 #include "bivariate_polynomial.h"
-#include "core/ggsw/ggsw.h"
+#include "core/ggsw/ggsw_arithmetic.h"
 #include "core/ggsw/glwegadget.h"
-#include "core/glwe/glwe.h"
+#include "core/glwe/glwe_arithmetic.h"
 #include "ggsw_params.h"
 #include "glwe_ciphertext.h"
 #include "glwe_key.h"
