+++ tests/unit/core/ggsw/test_glwegadget_half_product.c
@@ -5,8 +5,8 @@
 #include <stdio.h>
 
 #include "bivariate_polynomial.h"
-#include "core/ggsw/ggsw.h"
-#include "core/glwe/glwe.h"
+#include "core/ggsw/ggsw_arithmetic.h"
+#include "core/glwe/glwe_arithmetic.h"
 #include "core/glwe/glwe_ciphertext.h"
 #include "core/glwe/glwe_transform_key.h"
 #include "ggsw_params.h"
