+++ tests/unit/core/glwe/test_glwe.c
@@ -4,7 +4,7 @@
 
 #include "bivariate_polynomial.h"
 #include "common/spqlios_alias.h"
-#include "core/glwe/glwe.h"
+#include "core/glwe/glwe_arithmetic.h"
 #include "glwe_key.h"
 #include "glwe_params.h"
 #include "glwe_transform_key.h"
